scripts/requirements.txt
cairosvg 
pillow
